refactor: CommentRepository 삭제
