+++ Projects/TDDomain/Sources/UseCase/BlockCommentUseCase.swift
@@ -5,9 +5,9 @@ public protocol BlockCommentUseCase {
 }
 
 public final class BlockCommentUseCaseImpl: BlockCommentUseCase {
-    private let repository: CommentRepository
+    private let repository: SocialRepository
     
-    public init(repository: CommentRepository) {
+    public init(repository: SocialRepository) {
         self.repository = repository
     }
