+++ Projects/TDDomain/Sources/UseCase/FetchCommentUseCase.swift
@@ -7,9 +7,9 @@ public protocol FetchCommentUseCase {
 }
 
 public final class FetchCommentUseCaseImpl: FetchCommentUseCase {
-    private let repository: CommentRepository
+    private let repository: SocialRepository
     
-    public init(repository: CommentRepository) {
+    public init(repository: SocialRepository) {
         self.repository = repository
     }
