+++ Projects/TDDomain/Sources/UseCase/ReportCommentUseCase.swift
@@ -5,9 +5,9 @@ public protocol ReportCommentUseCase {
 }
 
 public final class ReportCommentUseCaseImpl: ReportCommentUseCase {
-    private let repository: CommentRepository
+    private let repository: SocialRepository
     
-    public init(repository: CommentRepository) {
+    public init(repository: SocialRepository) {
         self.repository = repository
     }
