+++ Projects/TDDomain/Sources/UseCase/ToggleCommentUseCase.swift
@@ -5,9 +5,9 @@ public protocol ToggleCommentLikeUseCase {
 }
 
 public final class ToggleCommentLikeUseCaseImpl: ToggleCommentLikeUseCase {
-    private let repository: CommentRepository
+    private let repository: SocialRepository
     
-    public init(repository: CommentRepository) {
+    public init(repository: SocialRepository) {
         self.repository = repository
     }
