+++ Projects/TDDomain/Sources/UseCase/UpdateCommentUseCase.swift
@@ -5,9 +5,9 @@ public protocol UpdateCommentUseCase {
 }
 
 public final class UpdateCommentUseCaseImpl: UpdateCommentUseCase {
-    private let repository: CommentRepository
+    private let repository: SocialRepository
     
-    public init(repository: CommentRepository) {
+    public init(repository: SocialRepository) {
         self.repository = repository
     }
